+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 CHANGES:
 
 * Added a new script iam/show-password-policy.py. ([@TGWolf][])
+* Added a new section for ssm scripts. ([@TGWolf][])
+* Added a new script ssm/list-all-parameters.py. ([@TGWolf][])
 
 ## 1.0.6 (November 22, 2018)